+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Bug #12: Clarify fill and stroke attribute documentation for SVG 2 compliance (@terabytesoftw)
 - Enh #13: Add `HasStrokeLineCap` trait and corresponding tests for managing SVG `stroke-linecap` attribute (@terabytesoftw)
 - Enh #14: Add `HasStrokeLineJoin` trait and corresponding tests for managing SVG `stroke-linejoin` attribute (@terabytesoftw)
+- Enh #15: Add `HasStrokeWidth` trait and corresponding tests for managing SVG `stroke-width` attribute (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024